+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Dart2JS
 
+* Removed `--no-defer-class-types` and `--no-new-deferred-split'.
+
 ### Tools
 
 #### Dartanalyzer